[refact/#2] api 파일 분리
